refactor: split syscontracts package from internal/oracle
Move syscontracts.go + deposit_contract.go (+ tests) from internal/oracle
to a new internal/syscontracts package. AddCanonicalSystemContracts,
DepositContractCode and DepositContractAddress are production code called
from main.go's --spec path; internal/oracle is left as the test-only
differential-output helpers (Reproduce / devkeys).

PR-review batch 3, item 14 (B-9): clearer "this is production" boundary;
0 LOC saved (pure file move + import rewrite). Builds clean; entitygen,
syscontracts, e2e_testing, and geth golden tests pass.
