+++ internal/syscontracts/syscontracts.go
@@ -1,4 +1,4 @@
-package oracle
+package syscontracts
 
 import (
 	"github.com/ethereum/go-ethereum/common"
