+++ internal/syscontracts/syscontracts_test.go
@@ -1,4 +1,4 @@
-package oracle
+package syscontracts
 
 import (
 	"bytes"
